finalize display module and controller shield
Display Module: Reordered the WS2813 LEDs to be consistent with normal
7-segment LED numbers (a-g). Changed schematics to be in the same
orientation as the board in top view. Re-adds the capacitors for the
WS2812 LEDs and the LDR line, which inadvertently got lost in the last
commit. Added silk screen.

Controller Shield: Added large electrolyte capacitor to buffer for the
WS2812 LEDs. Added a series resistor for the WS2812 data line. Added
silk screen and corrected wrong pinout of the I2C and serial header.