Fix communication between track and bridge
Added lots of try/catch handlers to avoid the bridge to throw ever
completely, as this will reset the serial port. Started to change
the highscore and leaderboard routines, as it may happen that a
heat must be restarted and so the cumulative scores and times must
be recalculated from all heat results. Added support to re-create
a missing or corrupt heat of a race.