Add local login functionality
Whenever the app gets freshly installed, a new JWT token secret is
generated as well as a new temporary root password. As long as it
is not changed it can be retrieved by an SPA app and used to get
an initial token with which a root passwd change can be started.